refactor: components file structure
