+++ src/components/custom/Checkbox.vue
@@ -9,7 +9,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import Icon from '@/components/Icon.vue'
+import Icon from '@/components/icons/Icon.vue'
 
 export default defineComponent({
     components: { Icon }
