+++ src/pages/Home.vue
@@ -11,8 +11,8 @@ import { defineComponent } from 'vue'
 import Header from '@/components/Header.vue'
 import Content from '@/components/Content.vue'
 import Footer from '@/components/Footer.vue'
-import Loading from '@/components/Loading.vue'
-import Toast from '@/components/Toast.vue'
+import Loading from '@/components/custom/Loading.vue'
+import Toast from '@/components/custom/Toast.vue'
 import { mapState } from 'vuex'
 
 export default defineComponent({
